Check for overflow when calculating buffer size
Use recently added check_mul_overflow() (added in f0907827a8a9152aedac of
the kernel).

Types of arguments passed to check_mul_overflow() must be the same so
change type of parameters to int and use int for bytes_per_line and
screen_size (even though it would make more sense for them to be u32 due
to type of struct members were values are stored). But since we check for
overflow we are guaranteed positive values.

Should perhaps modify width/height to be a single parameter and check that
it has a reasonable size in set callback, but leave as is for now.